Remove 'dataset' namespace property and support for public dataset interface
The public dataset interface is going to be turned off by ECMWF at the end of June 2021, with the ADS and CDS being the new public interfaces to this data.
There is therefore no longer a case to support this interface in this muninn plugin.
The only remaining supported interface for this plugin is the direct mars interface.