Allow specifying the packing approach
As of ifs 48r1 (to be released later this year) the default packing type for GRIB2 data from ECMWF will be CCSDS packing instead of simple packing.
To still allow retrieving data using the simple packing approach it will be needed to supply packing='simple' to the MARS request.
This commit makes this possible.